+++ CHANGES.md
@@ -4,7 +4,7 @@ Gene Stacker Changes
 Version 1.7 (under development)
 -------------------------------
 
- - [TODO] Colored output graphs. A new command line option `-nc,--no-color` is provided to produce greyscale graphs.
+ - Default colored output graphs. A new command line option `-nc,--no-color` is provided to produce greyscale graphs. **Note:** it is advised to install the latest version of Graphviz if colored graphs are not rendered correctly.
 
 Version 1.6 (10 September 2014)
 ------------------------------